Rename package contextpropagation to context
