+++ context-propagation-api/src/main/java/io/micrometer/context/ContextContainerAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.contextpropagation;
+package io.micrometer.context;
 
 import java.util.ServiceLoader;
