+++ context-propagation-api/src/main/java/io/micrometer/context/ContextScope.java
@@ -13,14 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.contextpropagation;
+package io.micrometer.context;
+
 
 /**
- * Storage for a {@link Namespace}.
- *
- * @author Marcin Grzejszczak
- * @since 1.0.0
+ * Demarcates the scope of restored ThreadLocal values.
  */
-public interface Store {
+public interface ContextScope extends AutoCloseable {
+
+    @Override
+    void close();
 
 }
