+++ context-propagation-api/src/test/java/io/micrometer/context/ObservationThreadLocalAccessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.contextpropagation;
+package io.micrometer.context;
 
 public class ObservationThreadLocalAccessor implements ThreadLocalAccessor {
